+++ NEWS.md
@@ -23,8 +23,10 @@
   columns to these functions will now compute the min/max/sum in each column 
   separately. Use `pl$min_horizontal()` `pl$max_horizontal()`, and 
   `pl$sum_horizontal()` instead for rowwise computation (#508).
-- `$is_not()` is deprecated and will be removed in 0.12.0. Use `$not_()` instead 
-  (#511).
+- `$is_not()` is deprecated and will be removed in 0.12.0. Use `$not()` instead 
+  (#511, #531).
+- `$is_first()` is deprecated and will be removed in 0.12.0. Use `$is_first_distinct()`
+  instead (#531).
 - In `pl$concat()`, the argument `to_supertypes` is removed. Use the suffix 
   `"_relaxed"` in the `how` argument to cast columns to their shared supertypes
   (#523).
@@ -56,6 +58,7 @@
   date/datetime/numeric indices (#470).
 - New methods `$name$to_lowercase()` and `$name$to_uppercase()` to transform 
   variable names (#529).
+- New method `$is_last_distinct()` (#531).
 - New methods of the Expressions class, `$floor_div()`, `$mod()`, `$eq_missing()`
   and `$neq_missing()`. The base R operators `%/%` and `%%` for Expressions are
   now translated to `$floor_div()` and `$mod()` (#523).